+++ mobile-backend-psql/src/main/kotlin/itmo/routes/RoomRouting.kt
@@ -69,14 +69,13 @@ fun Route.roomRouting() {
 
                 val notValid = entity.name.isBlank() || entity.userId <= 0
                 if (notValid) {
-                    throw BadRequestException("Некорректное значение одного или нескольких полей - name, typeId, userId")
+                    throw BadRequestException("Некорректное значение одного или нескольких полей - name, userId")
                 }
 
-                // TODO проверка UNIQUE у пользователя
                 log(
                     "POST /rooms",
                     "${entity.userId}",
-                    "insert room name ${entity.name}",
+                    "Добавлена комната с именем ${entity.name} пользователю #${entity.userId}",
                     "success"
                 )
 
@@ -97,14 +96,17 @@ fun Route.roomRouting() {
             val id = call.request.queryParameters["roomId"]?.toIntOrNull()
 
             try {
-
                 if (id == null || id <= 0) {
                     throw BadRequestException("Указан некорректный id=$id")
                 }
 
+                if (dao.findById(id) == null) {
+                    throw BadRequestException("Комната #$id не существует")
+                }
+
                 dao.deleteById(id)
                 log("DELETE /rooms?roomId=$id", "-1", "Комната #$id удалёна", "success")
-                call.respond(HttpStatusCode.OK, "Комната $id удалена")
+                call.respond(HttpStatusCode.OK, "Комната #$id удалена")
 
             } catch (e: BadRequestException) {
                 log(
